backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
Flask-Login==0.6.3
gunicorn==21.2.0
